requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=5.1=1_gnu
absl-py=1.4.0=pypi_0
antlr4-python3-runtime=4.9.3=pypi_0
appdirs=1.4.4=pypi_0
asttokens=2.2.1=pypi_0
attrs=22.2.0=pypi_0
audioread=3.0.0=pypi_0
backcall=0.2.0=pypi_0
beautifulsoup4=4.12.0=pypi_0
black=23.3.0=pypi_0
blas=1.0=mkl
boltons=23.0.0=py38h06a4308_0
bottleneck=1.3.7=pypi_0
brotlipy=0.7.0=py38h27cfd23_1003
bzip2=1.0.8=h7b6447c_0
c-ares=1.19.0=h5eee18b_0
ca-certificates=2023.01.10=h06a4308_0
cachetools=5.3.0=pypi_0
certifi=2022.12.7=py38h06a4308_0
cffi=1.15.1=py38h5eee18b_3
chainer=6.0.0=dev_0
charset-normalizer=2.0.4=pyhd3eb1b0_0
ci-sdr=0.0.2=pypi_0
click=8.1.3=pypi_0
cmake=3.22.1=h1fce559_0
colorama=0.4.6=pypi_0
comm=0.1.3=pypi_0
conda=23.3.1=py38h06a4308_0
conda-package-handling=2.0.2=py38h06a4308_0
conda-package-streaming=0.7.0=py38h06a4308_0
configargparse=1.5.3=pypi_0
contourpy=1.0.7=pypi_0
cryptography=39.0.1=py38h9ce1e76_0
ctc-segmentation=1.7.4=pypi_0
cudatoolkit=11.3.1=h2bc3f7f_2
cycler=0.11.0=pypi_0
cython=0.29.33=pypi_0
debugpy=1.6.6=pypi_0
decorator=5.1.1=pypi_0
distance=0.1.3=pypi_0
docker-pycreds=0.4.0=pypi_0
editdistance=0.5.2=pypi_0
einops=0.6.0=pypi_0
espnet=202301=dev_0
espnet-model-zoo=0.1.7=pypi_0
espnet-tts-frontend=0.0.3=pypi_0
executing=1.2.0=pypi_0
expat=2.4.9=h6a678d5_0
fairscale=0.4.13=pypi_0
fast-bss-eval=0.1.3=pypi_0
fastdtw=0.3.4=pypi_0
ffmpeg-python=0.2.0=pypi_0
filelock=3.10.7=pypi_0
flit-core=3.8.0=py38h06a4308_0
fonttools=4.39.3=pypi_0
future=0.18.3=pypi_0
g2p-en=2.1.0=pypi_0
gdown=4.7.1=pypi_0
gettext=0.21.0=h39681ba_1
gitdb=4.0.10=pypi_0
gitpython=3.1.31=pypi_0
google-auth=2.17.1=pypi_0
google-auth-oauthlib=1.0.0=pypi_0
grpcio=1.53.0=pypi_0
h5py=3.8.0=pypi_0
huggingface-hub=0.13.3=pypi_0
humanfriendly=10.0=pypi_0
hydra-core=1.3.2=pypi_0
icu=58.2=hf484d3e_1000
idna=3.4=py38h06a4308_0
importlib-metadata=4.13.0=pypi_0
importlib-resources=5.12.0=pypi_0
inflect=6.0.2=pypi_0
intel-openmp=2021.4.0=h06a4308_3561
ipykernel=6.22.0=pypi_0
ipython=8.12.0=pypi_0
jaconv=0.3.4=pypi_0
jamo=0.4.1=pypi_0
jedi=0.18.2=pypi_0
joblib=1.2.0=pypi_0
jsonpatch=1.32=pyhd3eb1b0_0
jsonpointer=2.1=pyhd3eb1b0_0
jsonschema=4.17.3=pypi_0
jupyter-client=8.1.0=pypi_0
jupyter-core=5.3.0=pypi_0
kaldiio=2.18.0=pypi_0
kiwisolver=1.4.4=pypi_0
krb5=1.19.4=h568e23c_0
ld_impl_linux-64=2.38=h1181459_1
libcurl=7.88.1=h91b91d3_0
libedit=3.1.20221030=h5eee18b_0
libev=4.33=h7f8727e_1
libffi=3.4.2=h6a678d5_6
libflac=1.3.4=h27087fc_0
libgcc-ng=11.2.0=h1234567_1
libgomp=11.2.0=h1234567_1
libnghttp2=1.46.0=hce63b2e_0
libogg=1.3.4=h7f98852_1
libopus=1.3.1=h7f98852_1
librosa=0.9.2=pypi_0
libsndfile=1.0.31=h9c3ff4c_1
libssh2=1.10.0=h8f2d780_0
libstdcxx-ng=11.2.0=h1234567_1
libuv=1.44.2=h5eee18b_0
libvorbis=1.3.7=h9c3ff4c_0
libxml2=2.10.3=hcbfbd50_0
llvmlite=0.39.1=pypi_0
lxml=4.9.2=pypi_0
lz4-c=1.9.4=h6a678d5_0
markdown=3.4.3=pypi_0
markupsafe=2.1.2=pypi_0
matplotlib=3.7.1=pypi_0
matplotlib-inline=0.1.6=pypi_0
mir-eval=0.7=pypi_0
mkl=2021.4.0=h06a4308_640
mkl-service=2.4.0=py38h7f8727e_0
mkl_fft=1.3.1=py38hd3c417c_0
mkl_random=1.2.2=py38h51133e4_0
morfessor=2.0.6=pypi_0
musdb=0.4.0=pypi_0
museval=0.4.0=pypi_0
mypy-extensions=1.0.0=pypi_0
nara-wpe=0.0.9=pypi_0
ncurses=6.4=h6a678d5_0
nest-asyncio=1.5.6=pypi_0
nltk=3.8.1=pypi_0
nnmnkwii=0.1.2=pypi_0
numba=0.56.4=pypi_0
numpy=1.23.5=py38h14f4228_0
numpy-base=1.23.5=py38h31eccc5_0
oauthlib=3.2.2=pypi_0
omegaconf=2.3.0=pypi_0
openssl=1.1.1t=h7f8727e_0
opt-einsum=3.3.0=pypi_0
packaging=23.0=py38h06a4308_0
pandas=1.5.3=pypi_0
parso=0.8.3=pypi_0
pathspec=0.11.1=pypi_0
pathtools=0.1.2=pypi_0
pexpect=4.8.0=pypi_0
pickleshare=0.7.5=pypi_0
pillow=9.4.0=pypi_0
pip=23.0.1=py38h06a4308_0
pkgutil-resolve-name=1.3.10=pypi_0
platformdirs=3.2.0=pypi_0
pluggy=1.0.0=py38h06a4308_1
pooch=1.7.0=pypi_0
portalocker=2.7.0=pypi_0
praatio=6.0.0=pypi_0
prompt-toolkit=3.0.38=pypi_0
protobuf=3.20.1=pypi_0
psutil=5.9.4=pypi_0
ptyprocess=0.7.0=pypi_0
pure-eval=0.2.2=pypi_0
pyaml=21.10.1=pypi_0
pyasn1=0.4.8=pypi_0
pyasn1-modules=0.2.8=pypi_0
pycosat=0.6.4=py38h5eee18b_0
pycparser=2.21=pyhd3eb1b0_0
pydantic=1.10.7=pypi_0
pygments=2.14.0=pypi_0
pyopenssl=23.0.0=py38h06a4308_0
pyparsing=3.0.9=pypi_0
pypinyin=0.44.0=pypi_0
pyrsistent=0.19.3=pypi_0
pysocks=1.7.1=py38h06a4308_0
pysptk=0.2.0=pypi_0
pystoi=0.3.3=pypi_0
python=3.8.16=h7a1cb2a_3
python-dateutil=2.8.2=pypi_0
pytorch=1.10.1=py3.8_cuda11.3_cudnn8.2.0_0
pytorch-mutex=1.0=cuda
pytorch-ranger=0.1.1=pypi_0
pytorch-wpe=0.0.1=pypi_0
pytz=2023.3=pypi_0
pyworld=0.3.2=pypi_0
pyyaml=6.0=pypi_0
pyzmq=25.0.2=pypi_0
readline=8.2=h5eee18b_0
regex=2023.3.23=pypi_0
requests=2.28.1=py38h06a4308_1
requests-oauthlib=1.3.1=pypi_0
resampy=0.4.2=pypi_0
rhash=1.4.1=h3c74f83_1
rsa=4.9=pypi_0
ruamel.yaml=0.17.21=py38h5eee18b_0
ruamel.yaml.clib=0.2.6=py38h5eee18b_1
sacrebleu=2.3.1=pypi_0
scikit-learn=1.2.2=pypi_0
scipy=1.10.1=pypi_0
sentencepiece=0.1.97=pypi_0
sentry-sdk=1.18.0=pypi_0
setproctitle=1.3.2=pypi_0
setuptools=65.6.3=py38h06a4308_0
simplejson=3.18.4=pypi_0
six=1.16.0=pyhd3eb1b0_1
smmap=5.0.0=pypi_0
soundfile=0.12.1=pypi_0
soupsieve=2.4=pypi_0
sqlite=3.41.1=h5eee18b_0
stack-data=0.6.2=pypi_0
stempeg=0.2.3=pypi_0
tabulate=0.9.0=pypi_0
tensorboard=2.12.1=pypi_0
tensorboard-data-server=0.7.0=pypi_0
tensorboard-plugin-wit=1.8.1=pypi_0
threadpoolctl=3.1.0=pypi_0
tk=8.6.12=h1ccaba5_0
tokenize-rt=5.0.0=pypi_0
tomli=2.0.1=pypi_0
toolz=0.12.0=py38h06a4308_0
torch-complex=0.4.3=pypi_0
torch-optimizer=0.3.0=pypi_0
torchaudio=0.10.1=py38_cu113
tornado=6.2=pypi_0
tqdm=4.65.0=py38hb070fc8_0
traitlets=5.9.0=pypi_0
typeguard=2.13.3=pypi_0
typing_extensions=4.4.0=py38h06a4308_0
unidecode=1.3.6=pypi_0
urllib3=1.26.15=py38h06a4308_0
wandb=0.14.0=pypi_0
wcwidth=0.2.6=pypi_0
werkzeug=2.2.3=pypi_0
wheel=0.38.4=py38h06a4308_0
xz=5.2.10=h5eee18b_1
youtube-dl=2021.12.17=pypi_0
zipp=3.15.0=pypi_0
zlib=1.2.13=h5eee18b_0
zstandard=0.19.0=py38h5eee18b_0
zstd=1.5.4=hc292b87_0
